requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==3.13.0
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==20.1.0
inflection==0.5.1
packaging==24.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
setuptools==72.1.0
sqlparse==0.5.1
uritemplate==4.1.1
wheel==0.43.0
whitenoise==6.4.0
